pyproject.toml
[project]
name = "speech-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gtts>=2.5.4",
    "loguru>=0.7.3",
    "playwright>=1.55.0",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "pygame>=2.6.1",
    "pyyaml>=6.0.2",
    "speechrecognition>=3.14.3",
]
